styling for intro pages
