+++ flutter/wtc/lib/intro_screens/intro_page1.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:lottie/lottie.dart';
 
 class IntroPage1 extends StatefulWidget {
   const IntroPage1({super.key});
@@ -11,9 +12,30 @@ class _IntroPage1State extends State<IntroPage1> {
   @override
   Widget build(BuildContext context) {
     return Container(
-      color: Colors.white,
+      color: const Color(0xffBD9F4C),
       alignment: Alignment.center,
-      child: const Text("Intro Page 1"),
+      child: Column(
+        children: [
+          const SizedBox(height: 150),
+          Lottie.asset("images/crops.json", height: 300),
+          const Text(
+            "Welcome to WTC",
+            style: TextStyle(
+              fontSize: 22,
+              fontWeight: FontWeight.bold,
+              color: Colors.black,
+            ),
+          ),
+          const SizedBox(height: 10),
+          const Text(
+            "Let's get you started.",
+            style: TextStyle(
+              fontSize: 18,
+              color: Colors.black,
+            ),
+          ),
+        ],
+      )
     );
   }
 }
