+++ flutter/wtc/lib/intro_screens/intro_page2.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+import 'package:lottie/lottie.dart';
 
 class IntroPage2 extends StatefulWidget {
   const IntroPage2({super.key});
@@ -11,14 +13,34 @@ class _IntroPage2State extends State<IntroPage2> {
   @override
   Widget build(BuildContext context) {
     return Container(
-      color: Colors.grey.shade300,
+      color: const Color(0xffBD9F4C),
       alignment: Alignment.center,
-      child: const Text(
-        "Intro Page 2",
-        style: TextStyle(
-          color: Colors.black,
-        ),
-      ),
+      child: Column(
+        children: [
+          const SizedBox(height: 150),
+          Lottie.asset("images/book.json", height: 300),
+          const Text(
+            "Get the information you want.",
+            style: TextStyle(
+              fontSize: 22,
+              fontWeight: FontWeight.bold,
+              color: Colors.black,
+            ),
+          ),
+          const SizedBox(height: 15),
+          const Padding(
+            padding: EdgeInsets.symmetric(horizontal: 20),
+            child: Text(
+              "Select the tags you're interested in and stay informed on those topics.",
+              style: TextStyle(
+                fontSize: 18,
+                color: Colors.black,
+              ),
+              textAlign: TextAlign.center,
+            ),
+          ),
+        ],
+      )
     );
   }
 }
