+++ flutter/wtc/pubspec.yaml
@@ -59,6 +59,7 @@ dependencies:
   flutter_launcher_icons: ^0.13.1
   web: ^0.5.1
   google_nav_bar: ^5.0.6
+  lottie: ^3.1.2
 
 
 dev_dependencies:
@@ -91,6 +92,8 @@ flutter:
     - images/google.png
     - images/apple.png
     - images/email.png
+    - images/crops.json
+    - images/book.json
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
